Word Counter
A program that counts word frequency. The program initially place every
word from a text file to a list. The program then builds builds a Map
from the contents of the List which is empirically timed. The program
reports the top "N" most frequently  used words from the text file which
is also empirically timed.